third_party/WebKit/LayoutTests/fast/backgrounds/size/parsing-background-size-values-expected.txt

This tests checks that all of the input values for background-size parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test("background-size: contain;") is "contain"
PASS test("background-size: cover;") is "cover"
PASS test("background-size: 100 100;") is ""
PASS test("background-size: 100px 100px;") is "100px 100px"
PASS test("background-size: auto 50px;") is "auto 50px"
PASS test("background-size: 50px auto;") is "50px"
PASS test("background-size: auto auto;") is "auto"
PASS test("background-size: 30% 20%;") is "30% 20%"
PASS test("background-size: 4em auto;") is "4em"
PASS test("background-size: 5em;") is "5em"
PASS test("-webkit-background-size: 5em ;") is "5em 5em"
PASS test("background-size: 100px 100px 100px;") is ""
PASS test("background-size: coconut;") is ""
PASS test("background-size: 100px,;") is ""
PASS test("background-size: 100px, 50%;") is "100px, 50%"
PASS test("-webkit-background-size: 100px, 50%;") is "100px, 50% 50%"
PASS test("background-size: 50% 100px, 2em, 100px 50%;") is "50% 100px, 2em, 100px 50%"
PASS test("-webkit-background-size: 50% 100px, 2em, 100px 50%;") is "50% 100px, 2em, 100px 50%"
PASS successfullyParsed is true

TEST COMPLETE
